+++ webview-ui/src/components/DatabaseConfiguration.tsx
@@ -1,14 +1,11 @@
 import { TextFieldType } from "@vscode/webview-ui-toolkit";
-import {
-  VSCodeDivider,
-  VSCodeTextField,
-  VSCodeDropdown,
-  VSCodeOption,
-  VSCodeButton,
-} from "@vscode/webview-ui-toolkit/react";
+import { VSCodeDivider, VSCodeTextField, VSCodeDropdown, VSCodeOption } from "@vscode/webview-ui-toolkit/react";
 import { DatabaseConnection } from "../../../src/configuration/data/DatabaseConnection";
 import { NO_PRE_CONFIGURED_DRIVER, ALL_DRIVERS } from "../../../src/configuration/drivers";
-
+import { useEffect, useState } from "react";
+import { vscodeApiWrapper } from "../utilities/vscodeApiWrapper";
+import { MessageData, MessageType } from "../../../src/configuration/transfer";
+import { UrlParts } from "../../../src/configuration/data/UrlParts";
 /**
  * Properties for creating a database configuration
  */
@@ -34,11 +31,145 @@ interface DatabaseConfigurationProps {
 }
 
 /**
+ * The new url values that should be updated.
+ *
+ * Only the fields given will be changed. If any field is `undefined`, then the old value will be used.
+ */
+interface NewUrlValues extends UrlParts {
+  /**
+   * The new database type.
+   */
+  databaseType?: string;
+}
+
+/**
  * Creates an DatabaseConfiguration input area. This can be used for the normal an the reference connection.
  * @param pProperties - the properties for creating the element
  * @returns the created element
  */
 export function DatabaseConfiguration(pProperties: DatabaseConfigurationProps) {
+  const [serverAddress, setServerAddress] = useState<string>("localhost");
+  const [port, setPort] = useState<number>(-1);
+  const [databaseName, setDatabaseName] = useState<string>("data");
+
+  /**
+   * Updates the url for the changed values of the url parts.
+   * @param newValues - the new values to set. You do not need to fill every value, just the new values
+   */
+  function updateUrl(newValues: NewUrlValues) {
+    const databaseType = newValues.databaseType ?? pProperties.databaseConnection?.databaseType ?? "";
+    const driver = ALL_DRIVERS.get(databaseType);
+    if (driver) {
+      // find out, if there are parameters in the old url
+      // these parameters need to be added to the new url as well
+      let parameters = "";
+      const oldUrl = pProperties.databaseConnection?.url;
+      if (oldUrl && oldUrl.includes("?")) {
+        parameters = oldUrl.substring(oldUrl.indexOf("?"));
+      }
+
+      const url = `${driver.jdbcName}${newValues.serverAddress ?? serverAddress}:${newValues.port ?? port}${
+        driver.separator
+      }${newValues.databaseName ?? databaseName}${parameters}`;
+
+      // set the new url
+      pProperties.onUpdate("url", url);
+    }
+  }
+
+  useEffect(() => updateAfterDatabaseConnectionChanged(), [pProperties.databaseConnection]);
+
+  /**
+   * Update server address, port and database name whenever database connection changes
+   */
+  function updateAfterDatabaseConnectionChanged(): void {
+    // extract the new url parts from the database connection
+    const extractedUrlParts = pProperties.databaseConnection?.extractUrlPartsFromDatabaseConfiguration();
+
+    // sets the new server address
+    const newServerAddress = extractedUrlParts?.serverAddress || "localhost";
+    setServerAddress(newServerAddress);
+
+    // sets the new port
+    const newPort = extractedUrlParts?.port || 0;
+    setPort(newPort);
+
+    // sets the new database name
+    const newDatabaseName = extractedUrlParts?.databaseName || "data";
+    setDatabaseName(newDatabaseName);
+
+    // updates the url with the new elements
+    updateUrl({
+      serverAddress: newServerAddress,
+      port: newPort,
+      databaseName: newDatabaseName,
+      databaseType: pProperties.databaseConnection?.databaseType,
+    });
+  }
+
+  /**
+   * Sets the new server address whenever a `onBlur` event was triggered.
+   * It updates automatically the url with the new url part.
+   * @param event - the `onBlur` event which has the new value
+   */
+  function handleServerAddress(event: React.FocusEvent<HTMLInputElement>): void {
+    const newServerAddress = event.target.value;
+    setServerAddress(newServerAddress);
+
+    updateUrl({ serverAddress: newServerAddress });
+  }
+
+  /**
+   * Sets the new port whenever a `onBlur` event was triggered.
+   * It updates automatically the url with the new url part.
+   * @param event - the `onBlur` event which has the new value
+   * @see https://github.com/microsoft/vscode-webview-ui-toolkit/issues/439:
+   *    Number inputs are not possible, therefore manually validation and parsing is necessary
+   */
+  function handlePort(event: React.FocusEvent<HTMLInputElement>): void {
+    const inputValue = event.target.value;
+
+    if (inputValue && /^\d*$/.test(inputValue)) {
+      // checks if the input value is a number
+
+      const newPort = parseInt(inputValue);
+      if (newPort >= 0 && newPort <= 65535) {
+        // checks if the port is in a valid range
+        setPort(newPort);
+
+        updateUrl({ port: newPort });
+      } else {
+        vscodeApiWrapper.postMessage(
+          new MessageData(MessageType.LOG_MESSAGE, {
+            level: "error",
+            message: `Port '${newPort}' needs to be a number between 0 and 65535`,
+            notifyUser: true,
+          })
+        );
+      }
+    } else {
+      vscodeApiWrapper.postMessage(
+        new MessageData(MessageType.LOG_MESSAGE, {
+          level: "error",
+          message: `Port '${inputValue}' needs to be a number`,
+          notifyUser: true,
+        })
+      );
+    }
+  }
+
+  /**
+   * Sets the new database name whenever a `onBlur` event was triggered.
+   * It updates automatically the url with the new url part.
+   * @param event - the `onBlur` event which has the new value
+   */
+  function handleDatabaseName(event: React.FocusEvent<HTMLInputElement>): void {
+    const newDatabaseName = event.target.value;
+    setDatabaseName(newDatabaseName);
+
+    updateUrl({ databaseName: newDatabaseName });
+  }
+
   return (
     <div>
       <fieldset>
@@ -47,6 +178,25 @@ export function DatabaseConfiguration(pProperties: DatabaseConfigurationProps) {
           <legend>Connection configuration</legend>
           {createInput(pProperties, "text", "username", "Username of the database")}
           {createInput(pProperties, "password", "password", "Password of the database")}
+        </fieldset>
+
+        <fieldset>
+          <legend>Database url</legend>
+
+          {pProperties.databaseConnection?.databaseType !== NO_PRE_CONFIGURED_DRIVER && (
+            <>
+              <VSCodeTextField value={serverAddress} onBlur={handleServerAddress}>
+                server address
+              </VSCodeTextField>
+              <VSCodeTextField value={port.toString()} onBlur={handlePort}>
+                port
+              </VSCodeTextField>
+              <VSCodeTextField value={databaseName} onBlur={handleDatabaseName}>
+                database name
+              </VSCodeTextField>
+            </>
+          )}
+
           {createInput(
             pProperties,
             "text",
@@ -54,17 +204,7 @@ export function DatabaseConfiguration(pProperties: DatabaseConfigurationProps) {
             "The url of the database",
             "For example: jdbc:mariadb://localhost:3306/data"
           )}
-          <VSCodeButton
-            disabled={pProperties.databaseConnection?.databaseType === NO_PRE_CONFIGURED_DRIVER}
-            formnovalidate={true}
-            onClick={() => generateBasicUrlForDriver(pProperties)}
-            appearance="secondary">
-            Generate basic url for selected database type
-            <span slot="start" className="codicon codicon-database"></span>
-          </VSCodeButton>
-        </fieldset>
-        <fieldset>
-          <legend>Database type</legend>
+
           <div className="dropdown-container">
             <label htmlFor="databaseTypeSelection">Database type for the configuration</label>
             <VSCodeDropdown
@@ -72,11 +212,13 @@ export function DatabaseConfiguration(pProperties: DatabaseConfigurationProps) {
               value={pProperties.databaseConnection?.databaseType}
               onInput={(e) => {
                 // @ts-expect-error error exists because type is not 100% correct. I cannot change the type and using any is against ESLint.
-                const value = e.target.value;
-                pProperties.onUpdate("databaseType", value);
+                const newDatabaseType = e.target.value;
+                pProperties.onUpdate("databaseType", newDatabaseType);
                 // remove classpath and driver values, when a pre-configured values was used
-                if (value !== NO_PRE_CONFIGURED_DRIVER) {
+                if (newDatabaseType !== NO_PRE_CONFIGURED_DRIVER) {
                   pProperties.onUpdate("driver", "");
+                  // updates the url, when a new database type was selected
+                  updateUrl({ databaseType: newDatabaseType, port: ALL_DRIVERS.get(newDatabaseType)?.port });
                 }
               }}>
               {createDatabaseSelections()}
@@ -154,16 +296,5 @@ export function DatabaseConfiguration(pProperties: DatabaseConfigurationProps) {
         {pLabel}
       </VSCodeTextField>
     );
-  }
-
-  function generateBasicUrlForDriver(pProperties: DatabaseConfigurationProps): void {
-    const databaseType = pProperties.databaseConnection?.databaseType;
-
-    if (databaseType && databaseType !== NO_PRE_CONFIGURED_DRIVER) {
-      const driver = ALL_DRIVERS.get(databaseType);
-      if (driver && pProperties.databaseConnection) {
-        pProperties.onUpdate("url", driver.basicUrlForConnecting);
-      }
-    }
   }
 }